move min, max, minmax, clamp to math functions #4425
