+++ algorithms/unit_tests/TestStdAlgorithmsMinMaxElementOps.cpp
@@ -43,7 +43,6 @@
 */
 
 #include <TestStdAlgorithmsCommon.hpp>
-#include <std_algorithms/Kokkos_MinMaxOperations.hpp>
 #include <std_algorithms/Kokkos_MinMaxElementOperations.hpp>
 
 namespace KE = Kokkos::Experimental;
