+++ core/src/Kokkos_MathematicalFunctions.hpp
@@ -346,6 +346,70 @@ KOKKOS_IMPL_MATH_UNARY_PREDICATE(isnan)
 #undef KOKKOS_IMPL_MATH_BINARY_FUNCTION
 
 }  // namespace Experimental
+
+// this fwd declaration is here because paid
+// is needed below and if we don't do this,
+// the compiler does not find it
+template <class T1, class T2>
+struct pair;
+
+namespace Experimental {
+
+// clamp
+template <class T>
+constexpr KOKKOS_INLINE_FUNCTION const T& clamp(const T& value, const T& lo,
+                                                const T& hi) {
+  KOKKOS_EXPECTS(!(hi < lo));
+  return (value < lo) ? lo : (hi < value) ? hi : value;
+}
+
+template <class T, class ComparatorType>
+constexpr KOKKOS_INLINE_FUNCTION const T& clamp(const T& value, const T& lo,
+                                                const T& hi,
+                                                ComparatorType comp) {
+  KOKKOS_EXPECTS(!comp(hi, lo));
+  return comp(value, lo) ? lo : comp(hi, value) ? hi : value;
+}
+
+// max
+template <class T>
+constexpr KOKKOS_INLINE_FUNCTION const T& max(const T& a, const T& b) {
+  return (a < b) ? b : a;
+}
+
+template <class T, class ComparatorType>
+constexpr KOKKOS_INLINE_FUNCTION const T& max(const T& a, const T& b,
+                                              ComparatorType comp) {
+  return comp(a, b) ? b : a;
+}
+
+// min
+template <class T>
+constexpr KOKKOS_INLINE_FUNCTION const T& min(const T& a, const T& b) {
+  return (b < a) ? b : a;
+}
+
+template <class T, class ComparatorType>
+constexpr KOKKOS_INLINE_FUNCTION const T& min(const T& a, const T& b,
+                                              ComparatorType comp) {
+  return comp(b, a) ? b : a;
+}
+
+// minmax
+template <class T>
+constexpr KOKKOS_INLINE_FUNCTION auto minmax(const T& a, const T& b) {
+  using return_t = ::Kokkos::pair<const T&, const T&>;
+  return (b < a) ? return_t{b, a} : return_t{a, b};
+}
+
+template <class T, class ComparatorType>
+constexpr KOKKOS_INLINE_FUNCTION auto minmax(const T& a, const T& b,
+                                             ComparatorType comp) {
+  using return_t = ::Kokkos::pair<const T&, const T&>;
+  return comp(b, a) ? return_t{b, a} : return_t{a, b};
+}
+
+}  // namespace Experimental
 }  // namespace Kokkos
 
 #endif
